requirements.txt
opencv-python
numpy
mtcnn
keras-facenet
tensorflow
scikit-learn
pymongo
matplotlib
fastapi
uvicorn
python-dotenv
cloudinary
python-multipart
ultralytics
retina-face
tf-keras
insightface
onnxruntime